Refactor map handling, improve debug printing, and adjust Makefile settings
- Modify `Makefile` to disable debugging and sanitization by default.
- Update `fill_and_replace.c` to improve tab replacement and memory handling.
	- Map is now a square without tabs.
- Change `int` to `size_t` for `rows` and `cols` in `structs.h` and related functions.
- Fix a logic error in `validate_map.c` when checking for wall boundaries.
- Improve debug output with `print_array_with_values` and `print_string_with_values`.
- Update `parse_validate_init_map.c` and `validate_file_and_init_data.c` to clean up validation logic.

+++ incl/structs.h
@@ -6,7 +6,7 @@
 /*   By: lade-kon <lade-kon@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/02/26 15:04:44 by lade-kon      #+#    #+#                 */
-/*   Updated: 2025/03/14 16:01:59 by lade-kon      ########   odam.nl         */
+/*   Updated: 2025/03/19 14:08:16 by lade-kon      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,8 +33,8 @@ typedef enum e_setting
 typedef struct s_map
 {
 	char	**map;
-	int		rows; //rows = if (ft_strlen(map[x][y] > rows) rows = y;
-	int		cols;
+	size_t	rows; //rows = if (ft_strlen(map[x][y] > rows) rows = y;
+	size_t	cols;
 }				t_map;
 
 typedef struct s_player
